Change website to use Hugo academic
This changes the gonum website to be based on hugo with the academic theme. This commit makes the basis for future additions to the site with the minimal tweaks to the theme. All changes were to the config.toml file and content folders, with the exception of a few changes to the theme. The first was making the gonum logo the thumbnail rather than the default, and the second was to change the header file in themes/academic/layouts/partials/header.html to add the meta tags for vanity imports. Lastly, this deletes the existing .git folder within the academic theme.

Otherwise, this moves the Intro to Gonum blog post (incorporating a few changes to make the markdown work), and adds a paper citing gonum to provide some minimal content.